+++ src/containers/AppContainer.ts
@@ -2,7 +2,7 @@ import { connect } from "react-redux";
 
 import { requestGetJKF, requestPutJKF, changeAutoSave } from "../actions";
 import { App } from "../components/App";
-import { AppState } from "../models";
+import { AppState } from "../store";
 
 const mapStateToProps = (state: AppState) => {
   return {
