+++ src/containers/KifuTreeContainer.ts
@@ -2,7 +2,7 @@ import { connect } from "react-redux";
 
 import { gotoPath, moveUpFork, moveDownFork, removeFork } from "../actions";
 import { KifuTreeComponent } from "../components/tree/KifuTree";
-import { KifuTreeState } from "../models";
+import { KifuTreeState } from "../store";
 
 const mapStateToProps = (state: KifuTreeState) => {
   return {
