+++ src/models/path.ts
@@ -1,0 +1,3 @@
+import { List } from "immutable";
+
+export type Path = List<number>;
