+++ src/models/sfen.ts
@@ -1,0 +1,1 @@
+export type SFEN = string;
